Edit information about cloning
If one tries to push to the lansingcodes repo
under the lansingcodes org, an error will occur,
as they do not have permissions to do so.
This change makes it clear that one should make a
fork in order to request changes.

Also, even if a person has permissions (say that
maybe the contributor role gives them this),
to make changes directly to the source repo is
dangerous and should be advised against (thus
forks are best here).